fix: resolve component test failure and consolidate mock locations (#197)
