+++ frontend/.gitignore
@@ -33,3 +33,4 @@ next-env.d.ts
 /blob-report/
 /playwright/.cache/
 .last-run.json
+/coverage
