+++ frontend/.prettierignore
@@ -1,2 +1,4 @@
 pnpm-lock.yaml
 package-lock.json
+coverage
+.next
