+++ frontend/lib/__mocks__/supabase.ts
